f64 throughout; 66 cast suppressions become 6
The crate computed in `f32` because genslip does, and every place the C stored a
`double` expression into a `float` was reproduced as a narrowing cast with an
`#[expect]` explaining it. That is 66 suppressions and a `let x = ...; x` shape
that existed only to hang them on. All of it is gone.

Sixty of the 66 disappear outright -- there is no narrowing left to suppress. The
six that remain are five in `units.rs`, where the reason is written once, and one
in the LCG. `units::exact` names "a subfault or sample count as a float" and
carries the 2^53 bound in its doc instead of twenty-five copies of the same
sentence; `units::samples` names the saturating float-to-int the C writes as
`(int)(t/dt + 0.5)` in eleven places, where Rust's cast maps NaN to zero and
saturates -- which is the behaviour we want and was previously nobody's decision.

Drift against genslip, worst over six cases:

  slip       2.2370e-06 -> 1.9317e-06     better by 14%
  onset_s    5.5640e-01 -> 5.5640e-01     unchanged; this is the solver, not rounding
  rake_deg   4.9993e-01 -> 4.9994e-01     unchanged; this is the format's quantum
  slip_rate  5.6171e-05 -> 7.9358e-05     worse by 41%

Both directions are the same mechanism and neither is a regression. The port has
stopped rounding where genslip rounds, so what the comparison now measures is
genslip's own single-precision error rather than the difference between two
roundings made at the same places. Slip is one scaled value and improves; a pulse
is normalised by a fold over its samples, where genslip accumulates more, so the
gap widens. Everything stays three to four orders inside `ACCEPTABLE`, which does
not move.

The port's own fields moved by ~1e-7 relative -- one f32 ulp -- across the board,
and the snapshot confirms the *logic* did not: draw count, silent-subfault count
and final seed are identical to the digit.

Cost: **+6.5%** on a whole rupture, 1.444 ms to 1.538 ms. Cheap for double
precision, and worth recording because the FFT's memory traffic doubled.

Three bugs my own sweep introduced, all found by the tests, all the same class:
removing `f64::from(...)` dropped the parentheses its argument needed.

  sqrt(a/n)               became  a / sqrt(n)
  0.6 * sin(0.5*arg)      became  0.3 * sin(arg)
  1 - (d-45)/(90-45)      became  1 - (d-45)/90 - 45

The last gave alpha_t = 1.325, outside the [1/1.1, 1] its definition allows, and
the calibration test said so by name. A syntax-aware rewrite would not have made
any of them -- `ast-grep --pattern 'f64::from($X)' --rewrite '($X)'` keeps the
grouping -- and a skill now says so.

Three test files needed judgement rather than a sweep. `float_identities.rs` stays
**f32**: its claims are about the original's arithmetic, not the port's, and
widening them would turn a record of why each rewrite was free into a test of
nothing. `source_parity.rs` and `rise_time_parity.rs` asserted bit equality with a
transcription and got it -- but only because the store to `float` rounded away the
difference between `cbrt` and `exp(ln x / 3)`, which `float_identities.rs` has
asserted is real the whole time. Both files were right about their own width; only
widening the port made the two statements meet. They now use a derived bound.

That bound is worth reading: the gap is *not* a couple of ulps. `ln(x)` is accurate
to `u` relative, so its absolute error is `|ln x| * u`, and exponentiating turns
that back into a relative error of `k * |ln x| * u`. On a moment of 1e26 that is
twenty times a naive ulp count -- which is exactly what measures, and what the
first attempt at the bound got wrong.

Python arrays are float64. `crates/srf` and `srf.py` stay float32, because an SRF
writes six significant figures and that is the format's resolution rather than a
shortcut; `assemble.py` is where the two meet. The harness keeps float32 where it
reproduces genslip's *arithmetic* -- the padded-extent rule, the GSF reader -- and
widens at the boundary, which is exact.

+++ crates/genslip/src/rng/mod.rs
@@ -49,9 +49,9 @@ pub trait DrawSource {
     /// A normal deviate with the given standard deviation and mean, advancing the
     /// stream.
     ///
-    /// `sigma` and `mean` are `f32` because that is what the fields carry; the
+    /// `sigma` and `mean` are `f64` because that is what the fields carry; the
     /// deviate is `f64`.
-    fn gaussian(&mut self, sigma: f32, mean: f32) -> f64;
+    fn gaussian(&mut self, sigma: f64, mean: f64) -> f64;
 
     /// Advance the stream by `count` normal deviates, discarding them.
     ///
